Move RawConverterUtils methods to RawConverter
