+++ explainability-connectors/src/test/java/org/kie/trustyai/connectors/kserve/v2/PayloadParserRawTest.java
@@ -24,7 +24,7 @@ class PayloadParserRawTest {
         final double value = random.nextDouble();
 
         ModelInferRequest.Builder builder = ModelInferRequest.newBuilder();
-        builder.addRawInputContents(RawConverterUtils.fromDouble(List.of(value)));
+        builder.addRawInputContents(RawConverter.fromDouble(List.of(value)));
 
         ModelInferRequest.InferInputTensor inputTensor = ModelInferRequest.InferInputTensor.newBuilder()
                 .setDatatype("FP64")
@@ -46,7 +46,7 @@ class PayloadParserRawTest {
         final List<Double> values = random.doubles(3).boxed().collect(Collectors.toList());
 
         ModelInferResponse.Builder response = ModelInferResponse.newBuilder();
-        response.addRawOutputContents(RawConverterUtils.fromDouble(values));
+        response.addRawOutputContents(RawConverter.fromDouble(values));
         ModelInferResponse.InferOutputTensor outputTensor = ModelInferResponse.InferOutputTensor.newBuilder()
                 .setDatatype("FP64")
                 .addShape(1).addShape(3).build();
@@ -66,7 +66,7 @@ class PayloadParserRawTest {
         final Random random = new Random();
         final List<Float> values = List.of(random.nextFloat(), random.nextFloat(), random.nextFloat());
         ModelInferResponse.Builder response = ModelInferResponse.newBuilder();
-        response.addRawOutputContents(RawConverterUtils.fromFloat(values));
+        response.addRawOutputContents(RawConverter.fromFloat(values));
         ModelInferResponse.InferOutputTensor outputTensor = ModelInferResponse.InferOutputTensor.newBuilder()
                 .setDatatype("FP32")
                 .addShape(1).addShape(3).build();
@@ -86,7 +86,7 @@ class PayloadParserRawTest {
         final Random random = new Random();
         final List<Double> values = List.of(random.nextDouble(), random.nextDouble(), random.nextDouble());
         ModelInferResponse.Builder response = ModelInferResponse.newBuilder();
-        response.addRawOutputContents(RawConverterUtils.fromDouble(values));
+        response.addRawOutputContents(RawConverter.fromDouble(values));
         ModelInferResponse.InferOutputTensor outputTensor = ModelInferResponse.InferOutputTensor.newBuilder()
                 .setDatatype("FP64")
                 .addShape(1).addShape(3).build();
@@ -106,7 +106,7 @@ class PayloadParserRawTest {
         final Random random = new Random();
         final List<Float> values = List.of(random.nextFloat(), random.nextFloat(), random.nextFloat());
         ModelInferRequest.Builder builder = ModelInferRequest.newBuilder();
-        builder.addRawInputContents(RawConverterUtils.fromFloat(values));
+        builder.addRawInputContents(RawConverter.fromFloat(values));
         ModelInferRequest.InferInputTensor tensor = ModelInferRequest.InferInputTensor.newBuilder()
                 .setDatatype("FP32")
                 .addShape(1).addShape(3).build();
@@ -126,7 +126,7 @@ class PayloadParserRawTest {
         final Random random = new Random();
         final List<Double> values = List.of(random.nextDouble(), random.nextDouble(), random.nextDouble());
         ModelInferRequest.Builder builder = ModelInferRequest.newBuilder();
-        builder.addRawInputContents(RawConverterUtils.fromDouble(values));
+        builder.addRawInputContents(RawConverter.fromDouble(values));
         ModelInferRequest.InferInputTensor tensor = ModelInferRequest.InferInputTensor.newBuilder()
                 .setDatatype("FP64")
                 .addShape(1).addShape(3).build();
