+++ explainability-connectors/src/test/java/org/kie/trustyai/connectors/kserve/v2/RawConverterTest.java
@@ -10,14 +10,13 @@ import org.junit.jupiter.api.Test;
 import com.google.protobuf.ByteString;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.kie.trustyai.connectors.kserve.v2.RawConverterUtils.*;
 
 class RawConverterTest {
 
     @Test
     void rawBoolean() {
         final List<Boolean> booleanList = List.of(true, false, true, false, true, false, true, false);
-        final ByteString byteString = fromBoolean(booleanList);
+        final ByteString byteString = RawConverter.fromBoolean(booleanList);
         final List<Boolean> booleanList2 = RawConverter.toBoolean(byteString);
         assertEquals(booleanList, booleanList2);
     }
@@ -25,7 +24,7 @@ class RawConverterTest {
     @Test
     void rawInteger() {
         final List<Integer> data = new Random().ints(20, 0, 100).boxed().collect(Collectors.toList());
-        final ByteString byteString = fromInteger(data);
+        final ByteString byteString = RawConverter.fromInteger(data);
         final List<Integer> converted = RawConverter.toInteger(byteString);
         assertEquals(data, converted);
     }
@@ -33,7 +32,7 @@ class RawConverterTest {
     @Test
     void rawLong() {
         final List<Long> data = new Random().longs(20, 0, 100).boxed().collect(Collectors.toList());
-        final ByteString byteString = fromLong(data);
+        final ByteString byteString = RawConverter.fromLong(data);
         final List<Long> converted = RawConverter.toLong(byteString);
         assertEquals(data, converted);
     }
@@ -46,7 +45,7 @@ class RawConverterTest {
             data.add(random.nextFloat());
         }
 
-        final ByteString byteString = fromFloat(data);
+        final ByteString byteString = RawConverter.fromFloat(data);
         final List<Float> converted = RawConverter.toFloat(byteString);
         assertEquals(data, converted);
     }
@@ -59,7 +58,7 @@ class RawConverterTest {
             data.add(random.nextDouble());
         }
 
-        final ByteString byteString = fromDouble(data);
+        final ByteString byteString = RawConverter.fromDouble(data);
         final List<Double> converted = RawConverter.toDouble(byteString);
         assertEquals(data, converted);
     }
